refactor: Head estructure and test
