+++ package.json
@@ -70,8 +70,6 @@
     "glob": "7.2.3",
     "husky": "7.0.4",
     "jest": "27.5.1",
-    "rosie": "2.1.0",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "1.15.6"
+    "rosie": "2.1.0"
   }
 }
