+++ src/components/Head/Head.jsx
@@ -9,7 +9,7 @@ function Head({ intl }) {
   return (
     <Helmet>
       <title>
-        {intl.formatMessage(messages['gradebook.page.title'], { siteName: getConfig().SITE_NAME })}
+        {intl.formatMessage(messages.PageTitle, { siteName: getConfig().SITE_NAME })}
       </title>
       <link rel="shortcut icon" href={getConfig().FAVICON_URL} type="image/x-icon" />
     </Helmet>
