+++ src/components/Head/index.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line no-restricted-exports
+export { default } from './Head';
