+++ src/components/Head/messages.js
@@ -1,9 +1,9 @@
 import { defineMessages } from '@edx/frontend-platform/i18n';
 
 const messages = defineMessages({
-  'gradebook.page.title': {
-    id: 'gradebook.page.title',
-    defaultMessage: 'Gradebook | {siteName}',
+  PageTitle: {
+    id: 'PageTitle',
+    defaultMessage: 'My Learner Records | {siteName}',
     description: 'Title tag',
   },
 });
