+++ src/components/Head/test/Head.test.jsx
@@ -1,16 +1,19 @@
 import React from 'react';
-import { IntlProvider } from '@edx/frontend-platform/i18n';
 import { Helmet } from 'react-helmet';
-import { mount } from 'enzyme';
 import { getConfig } from '@edx/frontend-platform';
-import Head from './Head';
+import { render, initializeMockApp } from '../../../setupTest';
+import Head from '../Head';
 
 describe('Head', () => {
+  beforeAll(async () => {
+    await initializeMockApp();
+  });
+
   const props = {};
   it('should match render title tag and favicon with the site configuration values', () => {
-    mount(<IntlProvider locale="en"><Head {...props} /></IntlProvider>);
+    render(<Head {...props} />);
     const helmet = Helmet.peek();
-    expect(helmet.title).toEqual(`Gradebook | ${getConfig().SITE_NAME}`);
+    expect(helmet.title).toEqual(`My Learner Records | ${getConfig().SITE_NAME}`);
     expect(helmet.linkTags[0].rel).toEqual('shortcut icon');
     expect(helmet.linkTags[0].href).toEqual(getConfig().FAVICON_URL);
   });
